test: split the generated-project suite by what it renders
tests/test_generate_project.py had grown to 1342 lines covering four
unrelated questions. It keeps the answers-shaped ones (slug, tracker
CLI, project kind, language, disambiguate, skills, the two rule
documents) and hands the rest to suites named for their subject:
workflows, commit-time hooks, and what the render reads from its
environment rather than from an answer.

The two helpers every suite needs — the render call and the file-content
assertion — move to tests/render_support.py and lose their leading
underscore, which said "private to this module" while four modules used
them. Same 43 tests, no assertion changed.

ADVANCES pandoscope/agentic-engineering-template#239.
